+++ wpull/testing/lua_hook_script.lua
@@ -1,5 +1,7 @@
 local injected_url_found = false
 
+wpull_hook.callbacks.version = 1
+
 wpull_hook.callbacks.engine_run = function()
   assert(wpull_hook.factory.get('Engine'))
   wpull_hook.factory.get('Engine').set_concurrent(2)